+++ src/OpenTelemetry.Exporter.Zipkin/CHANGELOG.md
@@ -8,7 +8,7 @@ Notes](../../RELEASENOTES.md).
 
 ## 1.15.0
 
-Released 2026-Jan-20
+Released 2026-Jan-21
 
 * This package is deprecated and it will stop receiving any updates in
   December 2026. Zipkin already supports OTLP ingestion. Change your exporter