+++ src/OpenTelemetry.Exporter.Zipkin/CHANGELOG.md
@@ -6,6 +6,9 @@ Notes](../../RELEASENOTES.md).
 
 ## Unreleased
 
+* Harden memory usage for endpoint caching and array tag serialization.
+  ([#7081](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7081))
+
 ## 1.15.2
 
 Released 2026-Apr-08