+++ src/OpenTelemetry.Exporter.Zipkin/CHANGELOG.md
@@ -8,7 +8,7 @@ Notes](../../RELEASENOTES.md).
 
 ## 1.15.3
 
-Released 2026-Apr-20
+Released 2026-Apr-21
 
 * Harden memory usage for endpoint caching and array tag serialization.
   ([#7081](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7081))